backend/pyproject.toml
[project]
name = "rfp-builder"
version = "0.1.0"
description = "Enterprise RFP Builder with Microsoft Agent Framework workflows"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tomli>=2.0.1",
    "agent-framework-core>=0.1.0",
    "openai>=1.12.0",
    "azure-identity>=1.15.0",
    "pypdf>=4.0.0",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.4",
    "Pillow>=10.0.0",
    "reportlab>=4.0.0",
    "pymmdc>=0.1.0",
    "aiofiles>=23.2.1",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
